# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change admin password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [dashboard]
      summary: Get revenue analytics (last 12 months)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Revenue analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"
  /dashboard/occupancy:
    get:
      operationId: getOccupancyAnalytics
      tags: [dashboard]
      summary: Get occupancy analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Occupancy data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccupancyAnalytics"
  /dashboard/recent-payments:
    get:
      operationId: getRecentPayments
      tags: [dashboard]
      summary: Get recent payments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /dashboard/dues-today:
    get:
      operationId: getDuesToday
      tags: [dashboard]
      summary: Get students with fees due today
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Students with dues today
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
  /dashboard/overdue:
    get:
      operationId: getOverdueStudents
      tags: [dashboard]
      summary: Get overdue fee summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Overdue students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
...
  /fees/mark-paid:
    post:
      operationId: markFeePaid
      tags: [fees]
      summary: Mark student fee as paid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkFeePaidInput"
      responses:
        "200":
          description: Fee marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /fees/mark-unpaid:
    post:
      operationId: markFeeUnpaid
      tags: [fees]
      summary: Mark student fee as unpaid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkFeeUnpaidInput"
      responses:
        "200":
          description: Fee marked unpaid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /reports/monthly-revenue:
    get:
      operationId: getMonthlyRevenueReport
      tags: [dashboard]
      summary: Monthly revenue report data
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRevenueReport"
  /reports/pending-fees:
    get:
      operationId: getPendingFeesReport
      tags: [dashboard]
      summary: Pending fees report
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending fees report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthResponse:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"
    Admin:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        whatsappNumber:
          type: ["string", "null"]
        libraryName:
          type: string
...
    Student:
      type: object
      required: [id, name, phone, fatherName, seatNumber, joiningDate, monthlyFee, feeDueDate, feeStatus]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        photoUrl:
          type: ["string", "null"]
        seatNumber:
          type: integer
        joiningDate:
          type: string
        monthlyFee:
          type: number
        feeDueDate:
          type: integer
        feeStatus:
          type: string
          enum: [paid, unpaid, overdue]
        nextDueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        whatsappNumber:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, studentId, amount, paymentDate, receiptNumber, month, year]
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: string
        seatNumber:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        receiptNumber:
          type: string
        month:
          type: integer
        year:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MarkFeePaidInput:
      type: object
      required: [studentId, amount, month, year]
      properties:
        studentId:
          type: string
        amount:
          type: number
        month:
          type: integer
        year:
          type: integer
        notes:
          type: string
    MarkFeeUnpaidInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: string
    DashboardStats:
      type: object
      required: [totalSeats, occupiedSeats, vacantSeats, totalStudents, pendingFees, todayCollection, monthlyCollection]
      properties:
        totalSeats:
          type: integer
        occupiedSeats:
          type: integer
        vacantSeats:
          type: integer
        totalStudents:
          type: integer
        pendingFees:
          type: number
        todayCollection:
          type: number
        monthlyCollection:
          type: number
        overdueCount:
          type: integer
    MonthlyRevenue:
      type: object
      required: [month, year, revenue, count]
      properties:
        month:
          type: integer
        year:
          type: integer
        monthLabel:
          type: string
        revenue:
          type: number
        count:
          type: integer
    OccupancyAnalytics:
      type: object
      required: [occupied, vacant, total]
      properties:
        occupied:
          type: integer
        vacant:
          type: integer
        total:
          type: integer
        occupancyRate:
          type: number
    MonthlyRevenueReport:
      type: object
      required: [month, year, totalRevenue, totalStudents, payments]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalRevenue:
          type: number
        totalStudents:
          type: integer
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"